feat: create book

+++ src/main/java/com/ea/libmanagement/domain/entities/BookAuthorId.java
@@ -1,12 +1,12 @@
 package com.ea.libmanagement.domain.entities;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
 import java.io.Serializable;
-import java.util.Objects;
 
 @Data
 @NoArgsConstructor
@@ -15,8 +15,8 @@ import java.util.Objects;
 public class BookAuthorId implements Serializable {
 
     @Column(name = "cd_book")
-    private Long cdBook;
+    private int cdBook;
 
     @Column(name = "cd_author")
-    private Long cdAuthor;
+    private int cdAuthor;
 }
